Adding functional tests for Operator
Fix nil pointer deref when err = nil

If err is = nil after getting the memcachedServerList, but the server
list is empty. We currently try to reference the err.Error() which
results in dereferencing a nil pointer. This change adds a string
describing the error to ensure a valid condition in either scenario.