Add startupProbe for Horizon pod
The Horizon pod is fairly frequently killed due to the current liveness
probe timing. This seems to happen more frequently during the initial
deployment while the system is under load. This change adds the
startup probe to provide some extra time for kubelet to validate the
functionality of the Horizon pod before terminating it.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>