Remove deprecated default_dashboard
The default_dashboard option is deprecated and the following warning
is shown.

WARNING:root:"dashboards" and "default_dashboard" in (local_)settings
is DEPRECATED now and may be unsupported in some future release.
The preferred way to specify the order of dashboards and the default
dashboard is the pluggable dashboard mechanism ... .