Use horizon deployment to test http response
This change uses the Horizon deployment pod to check
responsiveness of the Horizon webserver. This is necessary since we don't set
overrides for the endpointURL to create a valid Route. As such, we must check
via the Service. The main test here is to ensure Horizon is operational, so
from that perspective, this test should ensure basic http responsiveness of
the application during Kuttl tests.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>