Restrict ALLOWED_HOSTS
This change removes the ALLOWED_HOSTS=* and restricts it
instead to the Route Host and also the IP address of the pod.

To determine the IP address of the pod, this PR adds a function
that will retrieve the pods primary interface IP address and insert it
into the ALLOED_HOSTS list. This will facilitate liveness checks against
each of the pods.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>