Permit hostmount-anyuid for Horizon Service Account
This change adds the hostmount-anyuid permission to the horizon Service Account.
This will ensure that users are able to leverage NFS shares via the ExtraMounts
interface. Without this permission, we would need to set privileged: true on
the deployment, which ideally we would like to avoid.

This can be removed once we are no longer running with UID 0.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>