Add Customizing Horizon theme guide
This patch introduces a README with the instructions to deploy Horizon
with custom themes. It describes the basics behind the strategy to
provide theme assets to the Pods, and shows the steps required to
achieve this goal.
In addition, a quick deloyment helper is added, and the associated
sample based on the OpenStackControlPlane CR is included within the
existing samples/ directory.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>