Optimize memcache cache config when memcache instance is gone
Django seems to not detect/recover from the first memcached server
going away. When the memcache instance is gone, there are stack traces
in the horizon logs related to timeouts while trying to resolve the
dns name for the gone memcache pod. As a result horizon gets into
restart loop until that instance is back.

This change enables pooling for memcache. Also sets `ignore_exc` to
treated as a miss, instead of an exception.

Jira: OSPRH-16651

Co-outhored-by: Luca Miccini <lmiccini@redhat.com>
Signed-off-by: Martin Schuppert <mschuppert@redhat.com>